requirements.txt
steam
discord.py
